requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pyjwt
bcrypt
pydantic
